Organise and deduplicate fixtures
