+++ api/test/unit/conftest.py
@@ -12,7 +12,6 @@ from rest_framework.test import APIClient, APIRequestFactory
 import pook
 import pytest
 from elasticsearch import Elasticsearch
-from fakeredis import FakeRedis
 
 from api.models import (
     Audio,
@@ -38,19 +37,6 @@ from api.serializers.media_serializers import (
     MediaSearchRequestSerializer,
     MediaSerializer,
 )
-
-
-@pytest.fixture()
-def redis(monkeypatch) -> FakeRedis:
-    fake_redis = FakeRedis()
-
-    def get_redis_connection(*args, **kwargs):
-        return fake_redis
-
-    monkeypatch.setattr("django_redis.get_redis_connection", get_redis_connection)
-
-    yield fake_redis
-    fake_redis.client().close()
 
 
 @pytest.fixture
